refactor(Vsix): match namespace to "Commands" folder.
