+++ src/LibraryManager.Vsix/Commands/ManageLibrariesCommand.cs
@@ -12,7 +12,7 @@ using Microsoft.Web.LibraryManager.Vsix.Contracts;
 using Microsoft.Web.LibraryManager.Vsix.Shared;
 using Task = System.Threading.Tasks.Task;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Commands
 {
     internal sealed class ManageLibrariesCommand
     {
