+++ src/LibraryManager.Vsix/Commands/RestoreCommand.cs
@@ -9,7 +9,7 @@ using Microsoft.VisualStudio.Shell;
 using Microsoft.Web.LibraryManager.Vsix.Shared;
 using Task = System.Threading.Tasks.Task;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Commands
 {
     internal sealed class RestoreCommand
     {
