+++ src/LibraryManager.Vsix/Commands/RestoreOnBuildCommand.cs
@@ -15,7 +15,7 @@ using Microsoft.Web.LibraryManager.Vsix.Shared;
 using NuGet.VisualStudio;
 using Task = System.Threading.Tasks.Task;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Commands
 {
     internal sealed class RestoreOnBuildCommand
     {
